Moved model training into backend and trained models into models
