+++ backend/core/betai/models/models_train_moneyline.py
@@ -22,7 +22,11 @@ import nflreadpy as nfl
 # Setup
 # ============================================================
 
-MODEL_DIR = Path(__file__).resolve().parent / "trained_models"
+MODEL_DIR = (
+    Path(__file__).resolve().parent.parent.parent.parent
+    / "trained_models"
+    / "moneyline"
+)
 MODEL_DIR.mkdir(parents=True, exist_ok=True)
 
 
@@ -30,7 +34,8 @@ MODEL_DIR.mkdir(parents=True, exist_ok=True)
 # Data Loading and Feature Engineering
 # ============================================================
 
-def load_game_level_data(seasons=[2024, 2025]):
+
+def load_game_level_data(seasons=[2023, 2024, 2025]):
     """
     Load team stats and schedules, merge into one game-level dataset.
     Each row = 1 game, with home and away team stats joined.
@@ -44,23 +49,35 @@ def load_game_level_data(seasons=[2024, 2025]):
 
     # Keep only relevant columns from team stats
     keep_cols = [
-        "season", "week", "team",
-        "passing_epa", "rushing_epa",
-        "passing_yards", "rushing_yards",
-        "def_sacks", "def_interceptions", "def_fumbles_forced"
+        "season",
+        "week",
+        "team",
+        "passing_epa",
+        "rushing_epa",
+        "passing_yards",
+        "rushing_yards",
+        "def_sacks",
+        "def_interceptions",
+        "def_fumbles_forced",
+        "fg_pct",
+        "penalty_yards",
     ]
     stats = stats[keep_cols]
 
     # Rename for home and away before merging
-    home_stats = stats.rename(columns=lambda c: c if c in ["season", "week", "team"] else c + "_home")
-    away_stats = stats.rename(columns=lambda c: c if c in ["season", "week", "team"] else c + "_away")
+    home_stats = stats.rename(
+        columns=lambda c: c if c in ["season", "week", "team"] else c + "_home"
+    )
+    away_stats = stats.rename(
+        columns=lambda c: c if c in ["season", "week", "team"] else c + "_away"
+    )
 
     # Merge home stats
     df = schedules.merge(
         home_stats,
         left_on=["season", "week", "home_team"],
         right_on=["season", "week", "team"],
-        how="left"
+        how="left",
     )
 
     # Merge away stats
@@ -68,7 +85,7 @@ def load_game_level_data(seasons=[2024, 2025]):
         away_stats,
         left_on=["season", "week", "away_team"],
         right_on=["season", "week", "team"],
-        how="left"
+        how="left",
     )
 
     # Target variable: home win
@@ -87,8 +104,16 @@ def load_game_level_data(seasons=[2024, 2025]):
         "passing_yards_diff": ("passing_yards_home", "passing_yards_away"),
         "rushing_yards_diff": ("rushing_yards_home", "rushing_yards_away"),
         "sacks_diff": ("def_sacks_home", "def_sacks_away"),
-        "interceptions_diff": ("def_interceptions_home", "def_interceptions_away"),
-        "fumbles_forced_diff": ("def_fumbles_forced_home", "def_fumbles_forced_away"),
+        "interceptions_diff": (
+            "def_interceptions_home",
+            "def_interceptions_away",
+        ),
+        "fumbles_forced_diff": (
+            "def_fumbles_forced_home",
+            "def_fumbles_forced_away",
+        ),
+        "fg_pct_diff": ("fg_pct_home", "fg_pct_away"),
+        "penalty_yards_diff": ("penalty_yards_home", "penalty_yards_away"),
     }
 
     for new_col, (home_col, away_col) in diffs.items():
@@ -104,12 +129,12 @@ def load_game_level_data(seasons=[2024, 2025]):
     return df, features
 
 
-
 # ============================================================
 # Feature Selection
 # ============================================================
 
-def select_k_best(X, y, k=6):
+
+def select_k_best(X, y, k=10):
     """Select the top K features based on univariate F-test."""
     selector = SelectKBest(score_func=f_classif, k=min(k, X.shape[1]))
     X_new = selector.fit_transform(X, y)
@@ -121,6 +146,7 @@ def select_k_best(X, y, k=6):
 # ============================================================
 # Training Utilities
 # ============================================================
+
 
 def train_and_save(model, model_name, X_train, y_train, X_test, y_test):
     print(f"Training {model_name} ...")
@@ -146,18 +172,21 @@ def save_feature_list(feature_names, model_name):
 # Main Pipeline
 # ============================================================
 
+
 def main():
     # 1. Load and preprocess data
-    df, candidate_features = load_game_level_data(seasons=[2024, 2025])
+    df, candidate_features = load_game_level_data(seasons=[2023, 2024, 2025])
 
     X = df[candidate_features]
     y = df["home_win"]
 
     # 2. Feature selection
-    X_sel, selected_feats = select_k_best(X, y, k=min(6, len(candidate_features)))
+    X_sel, selected_feats = select_k_best(
+        X, y, k=min(9, len(candidate_features))
+    )
 
     # Save selected features
-    for name in ["logistic_regression", "naive_bayes", "random_forest"]:
+    for name in ["lr_moneyline", "nb_moneyline", "rf_moneyline"]:
         save_feature_list(selected_feats, name)
 
     # 3. Split train/test
@@ -167,19 +196,26 @@ def main():
 
     # 4. Train models
     accs = {}
-    accs["logistic_regression"] = train_and_save(
-        LogisticRegression(max_iter=500), "logistic_regression",
-        X_train, y_train, X_test, y_test
+    accs["lr_moneyline"] = train_and_save(
+        LogisticRegression(max_iter=500),
+        "lr_moneyline",
+        X_train,
+        y_train,
+        X_test,
+        y_test,
     )
 
-    accs["naive_bayes"] = train_and_save(
-        GaussianNB(), "naive_bayes",
-        X_train, y_train, X_test, y_test
+    accs["nb_moneyline"] = train_and_save(
+        GaussianNB(), "nb_moneyline", X_train, y_train, X_test, y_test
     )
 
-    accs["random_forest"] = train_and_save(
+    accs["rf_moneyline"] = train_and_save(
         RandomForestClassifier(n_estimators=100, random_state=42),
-        "random_forest", X_train, y_train, X_test, y_test
+        "rf_moneyline",
+        X_train,
+        y_train,
+        X_test,
+        y_test,
     )
 
     # 5. Summary
