+++ backend/trained_models/spread/nb_spread_features.txt
@@ -1,0 +1,8 @@
+passing_epa_diff
+rushing_epa_diff
+passing_yards_diff
+rushing_yards_diff
+sacks_diff
+interceptions_diff
+fumbles_forced_diff
+spread_line
