--- models/__init__.py
@@ -1,9 +1,0 @@
-from .logistic_regression import LogisticRegressionModel
-from .naive_bayes import NaiveBayesModel
-from .random_forest import RandomForestModel
-from .logistic_regression_moneyline import LRMoneyLine
-from .naive_bayes_moneyline import NBMoneyLine
-from .random_forest_moneyline import RFMoneyLine
-from .logistic_regression_spread import LRSpread
-from .naive_bayes_spread import NBSpread
-from .random_forest_spread import RFSpread
